Added BaseController node as Player/AI parent
I created BaseController to act as a parent node to the Player and AI nodes that will control respective characters.

This allows the encapsulation of the hierarchy of ascension tiers, as well as functionality that is stored by both controller types.

Also capitalized scene names for more formality and to improve presence in code.