Added Main Menu and Pause Menu
Set up functionality for the user to start at the main menu, begin a game or quit, and then pause the game and either resume or quit to the main menu.

Next I need to implement the kills/scoring aspect of the Player by connecting a "killed" signal to the player and displaying the proper kills/score in the HUD scoreLabel.